feat: add 'requestTimeout' field to AppOptions

+++ include/mach/AppOptions.hpp
@@ -7,8 +7,8 @@ namespace mach
 {
     struct AppOptions {
         std::string host = "127.0.0.1";
-        std::int32_t port = 3143;
-        std::int64_t threadCount = 4;
+        std::uint16_t port = 3143;
+        std::size_t threadCount = 4;
 
         std::chrono::seconds requestTimeout = std::chrono::seconds(30);
     };
